+++ data_test.go
@@ -2,17 +2,19 @@ package data
 
 import (
 	"fmt"
+	"math/rand"
 	"os"
 	"path/filepath"
+	"time"
 )
 
 var (
-	currentDir              string
-	jsonLoc, yamlLoc        string
-	rs                      []string
-	si, ssi, bi, ii, fi, vi Item
-	testItems               []Item
-	base                    *Vector
+	currentDir                              string
+	jsonLoc, yamlLoc                        string
+	rs                                      []string
+	si, ssi, bi, ii, ii64, ui, ui64, fi, vi Item
+	testItems                               []Item
+	base                                    *Vector
 )
 
 func testVector() *Vector {
@@ -32,7 +34,10 @@ func init() {
 	ssi = NewStringsItem("a.list", "a", "b", "c")
 	bi = NewBoolItem("a.bool", false)
 	ii = NewIntItem("a.int", 9)
-	fi = NewFloatItem("a.float", 9.9)
+	ii64 = NewInt64Item("a.int64", 999)
+	ui = NewUintItem("a.uint", 1)
+	ui64 = NewUint64Item("a.uint64", 111)
+	fi = NewFloat64Item("a.float", 9.9)
 	v := New("multi")
 	v.Set(
 		NewStringItem("vector.1", "ONE"),
@@ -44,7 +49,8 @@ func init() {
 	always2.Provide("the same")
 	ri := NewStringsItem("store.retrieval.string", rs...)
 	testItems = []Item{
-		ri, si, ssi, bi, ii, fi, vi, always1, always2,
+		ri, si, ssi, bi, ii, ii64, ui, ui64, fi, vi, always1, always2,
 	}
 	base = testVector().Clone()
+	rand.Seed(time.Now().UnixNano())
 }