+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## Changelog
 
-### Data 0.0.1 (19.09.2016)
+### data 0.0.3 (09.04.2019)
+
+- removal of internal error for xrr package
+
+
+### data 0.0.2 (29.06.2018)
+
+- Item added for tracking data of varied types
+- Trie added for managing Items
+- Vector added as primary management of trie held items
+- Store added as mechanism to manage vectors to varied other formats (json, yaml, etc.) 
+- testing for Item, Trie, Vector, and Store types
+
+
+### data 0.0.1 (19.09.2016)
 
 - initialize public package